+++ packages/backend/src/nest/connections-manager/connections-manager.module.ts
@@ -5,9 +5,10 @@ import { SocketModule } from '../socket/socket.module'
 import { StorageModule } from '../storage/storage.module'
 import { TorModule } from '../tor/tor.module'
 import { ConnectionsManagerService } from './connections-manager.service'
+import { StorageServiceClientModule } from '../storageServiceClient/storageServiceClient.module'
 
 @Module({
-  imports: [RegistrationModule, StorageModule, TorModule, SocketModule, LocalDbModule],
+  imports: [RegistrationModule, StorageModule, TorModule, SocketModule, LocalDbModule, StorageServiceClientModule],
   providers: [ConnectionsManagerService],
   exports: [ConnectionsManagerService],
 })