+++ packages/desktop/src/renderer/components/CreateJoinCommunity/JoinCommunity/JoinCommunity.stories.tsx
@@ -11,6 +11,7 @@ const Template: ComponentStory<typeof PerformCommunityActionComponent> = args =>
 }
 
 export const Component = Template.bind({})
+export const ServerError = Template.bind({})
 
 const args: PerformCommunityActionProps = {
   open: true,