+++ packages/e2e-tests/src/types.ts
@@ -1,0 +1,7 @@
+import { App } from './selectors'
+
+export interface UserTestData {
+  username: string
+  app: App
+  messages: string[]
+}