+++ packages/mobile/src/store/init/deepLink/deepLink.saga.test.ts
@@ -3,52 +3,29 @@ import { combineReducers } from '@reduxjs/toolkit'
 import { reducers } from '../../root.reducer'
 import { Store } from '../../store.types'
 import { prepareStore } from '../../../tests/utils/prepareStore'
-import { communities, connection, getInvitationCodes, identity } from '@quiet/state-manager'
+import { communities, getFactory } from '@quiet/state-manager'
 import { initActions } from '../init.slice'
 import { navigationActions } from '../../navigation/navigation.slice'
 import { ScreenNames } from '../../../const/ScreenNames.enum'
 import { deepLinkSaga } from './deepLink.saga'
-import { type Community, CommunityOwnership, type Identity, InvitationData, InvitationDataVersion } from '@quiet/types'
-import {
-  composeInvitationShareUrl,
-  validInvitationCodeTestData,
-  getValidInvitationUrlTestData,
-  validInvitationDatav1,
-} from '@quiet/common'
+import { type Community, CommunityOwnership, InvitationData, CreateNetworkPayload } from '@quiet/types'
+import { composeInvitationShareUrl, getValidInvitationUrlTestData, validInvitationDatav1 } from '@quiet/common'
+import { FactoryGirl } from 'factory-girl'
 
 describe('deepLinkSaga', () => {
   let store: Store
-
+  let factory: FactoryGirl
   const { code } = getValidInvitationUrlTestData(validInvitationDatav1[0])
 
   const validCode = code()
   const validData = validInvitationDatav1[0]
 
   const id = '00d045ab'
-
-  const community: Community = {
-    id,
-    name: '',
-    CA: {
-      rootCertString: '',
-      rootKeyString: '',
-    },
-    rootCa: '',
-    peerList: [],
-    onionAddress: '',
-    ownerCertificate: '',
-  }
-
-  const _identity: Partial<Identity> = {
-    id,
-    nickname: '',
-    userCsr: null,
-    userCertificate: null,
-    joinTimestamp: 0,
-  }
+  let community: Community
 
   beforeEach(async () => {
     store = (await prepareStore()).store
+    factory = await getFactory(store)
   })
 
   test('joins community', async () => {
@@ -58,17 +35,16 @@ describe('deepLinkSaga', () => {
         socketIOSecret: 'secret',
       })
     )
+    const createNetworkPayload: CreateNetworkPayload = {
+      ownership: CommunityOwnership.User,
+      inviteData: validData,
+    }
     const reducer = combineReducers(reducers)
     await expectSaga(deepLinkSaga, initActions.deepLink(validCode))
       .withReducer(reducer)
       .withState(store.getState())
       .put(initActions.resetDeepLink())
-      .put(
-        communities.actions.joinNetwork({
-          version: InvitationDataVersion.v1,
-          ...validData,
-        })
-      )
+      .put(communities.actions.createNetwork(createNetworkPayload))
       .put(
         navigationActions.replaceScreen({
           screen: ScreenNames.UsernameRegistrationScreen,
@@ -78,6 +54,10 @@ describe('deepLinkSaga', () => {
   })
 
   test('displays error if user already belongs to a community', async () => {
+    community = await factory.create<ReturnType<typeof communities.actions.addNewCommunity>['payload']>('Community', {
+      id,
+      name: 'rockets',
+    })
     store.dispatch(
       initActions.setWebsocketConnected({
         dataPort: 5001,
@@ -85,13 +65,7 @@ describe('deepLinkSaga', () => {
       })
     )
 
-    store.dispatch(
-      communities.actions.addNewCommunity({
-        ...community,
-        name: 'rockets',
-      })
-    )
-
+    store.dispatch(communities.actions.addNewCommunity(community))
     store.dispatch(communities.actions.setCurrentCommunity(community.id))
 
     const reducer = combineReducers(reducers)
@@ -109,15 +83,18 @@ describe('deepLinkSaga', () => {
       .not.put(
         communities.actions.createNetwork({
           ownership: CommunityOwnership.User,
-          peers: validData.pairs,
-          psk: validData.psk,
-          ownerOrbitDbIdentity: validData.ownerOrbitDbIdentity,
+          inviteData: validData,
         })
       )
       .run()
   })
 
   test("doesn't display error if user is connecting with the same community", async () => {
+    community = await factory.create<ReturnType<typeof communities.actions.addNewCommunity>['payload']>('Community', {
+      id,
+      name: '',
+      psk: validData.psk,
+    })
     store.dispatch(
       initActions.setWebsocketConnected({
         dataPort: 5001,
@@ -125,7 +102,10 @@ describe('deepLinkSaga', () => {
       })
     )
 
-    community.psk = validData.psk
+    const createNetworkPayload: CreateNetworkPayload = {
+      ownership: CommunityOwnership.User,
+      inviteData: validData,
+    }
 
     store.dispatch(communities.actions.addNewCommunity(community))
 
@@ -149,11 +129,8 @@ describe('deepLinkSaga', () => {
       })
       .put.like({
         action: {
-          type: communities.actions.joinNetwork.type,
-          payload: {
-            version: InvitationDataVersion.v1,
-            ...validData,
-          },
+          type: communities.actions.createNetwork.type,
+          payload: createNetworkPayload,
         },
       })
       .run()
@@ -169,6 +146,10 @@ describe('deepLinkSaga', () => {
       ],
       psk: 'BNlxfE=',
       ownerOrbitDbIdentity: 'testId',
+    }
+    const createNetworkPayload: CreateNetworkPayload = {
+      ownership: CommunityOwnership.User,
+      inviteData: invalidData,
     }
     const invalidCode = composeInvitationShareUrl(invalidData)
     store.dispatch(
@@ -193,12 +174,7 @@ describe('deepLinkSaga', () => {
           },
         },
       })
-      .not.put(
-        communities.actions.joinNetwork({
-          version: InvitationDataVersion.v1,
-          ...validData,
-        })
-      )
+      .not.put(communities.actions.createNetwork(createNetworkPayload))
       .run()
   })
 })